Move ak_from_json_new into place as ak_from_json.
